requirements.txt
psycopg2-binary==2.8.4
python-dotenv==0.13.0
